Create README.md
Sentiment analysis is the process of understanding and classifying emotions (positive, negative, and neutral) contained in writing using text analysis techniques. Sentiment analysis is often referred to as opinion mining. Where we will explore what emotions are behind every person's words.
The data that will be analyzed in this project are tweets crawled from twitter with a total of 1000 data with the keywords: vaccine, #vaccine and #vaccinecovid19. Dataset is data that does not yet have a label so that the processing uses unsupervised learning.